[#172938609] Attempts to cache identifier against MAC address.
The MAC address that we are able to access seems to rotate between each
connection on some devices. On other devices it seems to only rotate
every 15 minutes. This change will only really be useful for the latter
category of devices, but should not have an adverse impact on the
former.

With the 15 minute window there is a small chance that some readings
will be captured against the "wrong" cryptogram around midnight when it
rotates. However this should only be at most 15 minutes. If we end up
reducing the rotation period we should revisit this.